requirements.txt
fastapi
uvicorn
requests
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-prometheus
prometheus_client
